ensure cancellation safety
Let `event_listener` holds the `Arc`s of the `LocalMrInner`s
that are being used by RDMA ops to ensure cancellation safety.

`LocalMr` was replaced with New struct `LocalMrInner`.
Because every struct that can use APIs should hold an `Arc`
of the mr's metadata, but previous `LocalMr` can't hold itself.

Add `MrState` to avoid potential race condition

When the RDMA operations future were canceled but the related
mr is still being used by netdev or kernel. This is a potential
race condition. So we add `MrState` to `LocalMrInner`, to make
sure we can get slice or mutable slice only if the mr is readable
or writeable.

Fixes: #47 #56